Fix toasts
