+++ src/Toasts/components/Toast/Toast.styled.js
@@ -1,35 +1,4 @@
 import styled from 'styled-components';
-import * as C from '../../constants';
-
-const getColor = (theme, type) => {
-  switch (type) {
-    case C.TYPE_SUCCESS:
-      return {
-        border: theme.green700,
-        background: theme.green100,
-      };
-    case C.TYPE_INFORMATION:
-      return {
-        border: theme.blue700,
-        background: theme.blue100,
-      };
-    case C.TYPE_ERROR:
-      return {
-        border: theme.ruby800,
-        background: theme.ruby100,
-      };
-    case C.TYPE_WARNING:
-      return {
-        border: theme.gold900,
-        background: theme.gold50,
-      };
-    default:
-      return {
-        border: theme.blue800,
-        background: theme.blue100,
-      };
-  }
-};
 
 export const Message = styled.div`
     word-break: break-word;  
@@ -59,9 +28,7 @@ export const Toast = styled.div`
     }
 
     .icon {
-      background: ${({ theme, type }) => getColor(theme, type).border};
       fill: ${({ theme }) => theme.white};
-      border: ${({ theme, type }) => getColor(theme, type).border};
       border-radius: 100%;
       padding: 3px;
     }
@@ -83,7 +50,6 @@ export const Bar = styled.div`
     right: 0;
     display: block;
     height: .4rem;
-    background-color: ${({ theme, type }) => getColor(theme, type).border};
     width: ${({ done }) => `${100 - done}%`};
     transition: width 100ms;
     border-top-right-radius: 5px;
